requirements.txt
evdev
lifx-sdk